Need to add interactions=1 for distinct peaks.
Assuming interactions=1 is the same peak only works when peaks are
disconnected.  Once the density threshold is reduced so former peaks
are not subsumed into a larger density peak, we can have interactions=1
found where we lost the peak going from high->low and just discovered a
nearby lower density peak.

We can check on the peak_id, but if that is not the same, then we will
keep the peaks distinct.